requirements.txt
numpy
olefile-0.46
